@@ -1,20 +1,22 @@
 package lt.boldadmin.nexus.backend.test.unit.handler.worklog.duration
 
-import com.nhaarman.mockito_kotlin.doReturn
+import com.nhaarman.mockitokotlin2.doReturn
 import lt.boldadmin.nexus.api.service.worklog.duration.WorklogDurationService
+import lt.boldadmin.nexus.api.type.valueobject.DateRange
 import lt.boldadmin.nexus.backend.handler.worklog.duration.WorklogDurationHandler
 import lt.boldadmin.nexus.backend.route.Routes
 import lt.boldadmin.nexus.backend.test.unit.handler.create
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
 import org.mockito.Mockito.lenient
-import org.mockito.junit.MockitoJUnitRunner
+import org.mockito.junit.jupiter.MockitoExtension
 import org.springframework.test.web.reactive.server.WebTestClient
-import kotlin.test.assertEquals
+import java.time.LocalDate
 
-@RunWith(MockitoJUnitRunner::class)
+@ExtendWith(MockitoExtension::class)
 class WorklogDurationHandlerTest {
 
     @Mock
@@ -22,7 +24,7 @@ class WorklogDurationHandlerTest {
 
     private lateinit var webClient: WebTestClient
 
-    @Before
+    @BeforeEach
     fun setUp() {
         val contextStub = create()
         lenient()
@@ -82,4 +84,45 @@ class WorklogDurationHandlerTest {
 
         assertEquals(durationsSum, response.responseBody)
     }
+
+    @Test
+    fun `Sums work durations by project id and date range`() {
+        val projectId = "projectId"
+        val expectedDurationsSum = 123L
+        val dateRange = DateRange(LocalDate.of(2019, 5, 10), LocalDate.of(2019, 5, 15))
+        doReturn(expectedDurationsSum)
+            .`when`(worklogDurationServiceStub)
+            .sumWorkDurationsByProjectId(projectId, dateRange)
+
+        val response = webClient.get()
+            .uri("/worklog/project/$projectId/start/2019-05-10/end/2019-05-15/durations-sum")
+            .exchange()
+            .expectStatus()
+            .isOk
+            .expectBody(Long::class.java)
+            .returnResult()
+
+        assertEquals(expectedDurationsSum, response.responseBody!!)
+    }
+
+    @Test
+    fun `Sums work durations by collaborator id and date range`() {
+        val collaboratorId = "collaboratorId"
+        val expectedDurationsSum = 123L
+        val dateRange = DateRange(LocalDate.of(2019, 5, 10), LocalDate.of(2019, 5, 15))
+        doReturn(expectedDurationsSum)
+            .`when`(worklogDurationServiceStub)
+            .sumWorkDurationsByCollaboratorId(collaboratorId, dateRange)
+
+        val response = webClient.get()
+            .uri("/worklog/collaborator/$collaboratorId/start/2019-05-10/end/2019-05-15/durations-sum")
+            .exchange()
+            .expectStatus()
+            .isOk
+            .expectBody(Long::class.java)
+            .returnResult()
+
+        assertEquals(expectedDurationsSum, response.responseBody!!)
+    }
+
 }
